refactoring of NodeAppearance and EdgeAppearance into a single class that relies on the VisualPropertyType enum instead of individual methods
git-svn-id: http://chianti.ucsd.edu/svn@10279 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/trunk/src/cytoscape/visual/ui/icon/LineTypeIcon.java
@@ -103,8 +103,7 @@ public class LineTypeIcon extends VisualPropertyIcon {
 	public LineTypeIcon(BasicStroke stroke, int width, int height, String name, Color color) {
 		super(null, width, height, name, color);
 
-		final float lineWidth = (Float) VisualPropertyType.EDGE_LINE_WIDTH.getDefault(Cytoscape.getVisualMappingManager()
-		                                                                                       .getVisualStyle());
+		final float lineWidth = ((Double) VisualPropertyType.EDGE_LINE_WIDTH.getDefault(Cytoscape.getVisualMappingManager().getVisualStyle())).floatValue();
 
 		/*
 		 * Define a stroke for the line segment icon
